requirements.txt
discord.py~=2.6.3
python-dotenv~=1.0.0
aiohttp~=3.12.15